+++ scaffan/annotation.py
@@ -261,7 +261,7 @@ def read_annotations_ndpa(pth) -> list:
 
 
 def plot_annotations(
-    annotations, x_key="x", y_key="y", in_region=False, factor=[1, 1], show_id=True
+    annotations, x_key="x", y_key="y", in_region=False, factor=[1, 1], show_id=True, fontsize="x-small"
 ):
     if type(annotations) is dict:
         annotations = [annotations]
@@ -284,7 +284,7 @@ def plot_annotations(
                     np.min(y),
                     str(i),
                     c=annotation["color"],
-                    fontsize="x-small",
+                    fontsize=fontsize,
                 )